refactor(explorer): remove duplicate compontent
